fix: コードの可読性を上げるために、コードを細分化
